Fix: the issue with the sound
There was `auto sound = assets().getSound(soundName);` which created
a copy instead of referring to the initialized object in m_soundMap.
A reason of incorrect call was the declaration of method with consts
`const sf::Sound& getSound(const std::string& soundName) const;`.
This should not be, because the object `sf::Sound` should be mutable.